+++ src/main/java/com/ujcms/cms/core/domain/base/AttachmentBase.java
@@ -24,21 +24,21 @@ public class AttachmentBase implements Serializable {
      */
     @NotNull
     @Schema(description="附件ID")
-    private Integer id = 0;
+    private Long id = 0L;
 
     /**
      * 站点ID
      */
     @NotNull
     @Schema(description="站点ID")
-    private Integer siteId = 0;
+    private Long siteId = 0L;
 
     /**
      * 用户ID
      */
     @NotNull
     @Schema(description="用户ID")
-    private Integer userId = 0;
+    private Long userId = 0L;
 
     /**
      * 文件名称
@@ -85,27 +85,27 @@ public class AttachmentBase implements Serializable {
     @Schema(description="是否使用")
     private Boolean used = false;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getSiteId() {
+    public Long getSiteId() {
         return siteId;
     }
 
-    public void setSiteId(Integer siteId) {
+    public void setSiteId(Long siteId) {
         this.siteId = siteId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
