+++ src/main/java/com/ujcms/cms/core/domain/base/PerformanceTypeBase.java
@@ -23,14 +23,14 @@ public class PerformanceTypeBase implements Serializable {
      */
     @NotNull
     @Schema(description="绩效类型ID")
-    private Integer id = 0;
+    private Long id = 0L;
 
     /**
      * 站点ID
      */
     @NotNull
     @Schema(description="站点ID")
-    private Integer siteId = 0;
+    private Long siteId = 0L;
 
     /**
      * 名称
@@ -54,19 +54,19 @@ public class PerformanceTypeBase implements Serializable {
     @Schema(description="排序")
     private Long order = 0L;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getSiteId() {
+    public Long getSiteId() {
         return siteId;
     }
 
-    public void setSiteId(Integer siteId) {
+    public void setSiteId(Long siteId) {
         this.siteId = siteId;
     }
