+++ src/main/java/com/ujcms/cms/core/domain/base/SensitiveWordBase.java
@@ -23,7 +23,7 @@ public class SensitiveWordBase implements Serializable {
      */
     @NotNull
     @Schema(description="敏感词ID")
-    private Integer id = 0;
+    private Long id = 0L;
 
     /**
      * 名称
@@ -40,11 +40,11 @@ public class SensitiveWordBase implements Serializable {
     @Schema(description="是否启用")
     private Boolean enabled = true;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
